requirements.txt
llama-cpp-python,
requests>=2.31.0,
beautifulsoup4>=4.9.3